Modificada la organizacion de paquetes para que respete el diagrama
