+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/data/external/artist/artists/ArtistExternalService.kt
@@ -1,0 +1,9 @@
+package ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.artists
+
+
+import ayds.lisboa.songinfo.moredetails.fulllogic.domain.Artist
+
+interface ArtistExternalService {
+
+    fun getArtistFromLastFMAPI(artistName: String): Artist.ArtistData?
+}
