+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/data/external/artist/artists/ArtistExternalServiceImpl.kt
@@ -1,14 +1,14 @@
-package ayds.lisboa.songinfo.moredetails.fulllogic.model.entities.repository.externalArtist.artists
+package ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.artists
 
 import ayds.lisboa.songinfo.moredetails.fulllogic.LastFMAPI
-import ayds.lisboa.songinfo.moredetails.fulllogic.model.entities.Artist
-import ayds.lisboa.songinfo.moredetails.fulllogic.model.entities.repository.externalArtist.ArtistExternalService
+import ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.LastFMToArtistResolver
+import ayds.lisboa.songinfo.moredetails.fulllogic.domain.Artist
 
 
 internal class ArtistExternalServiceImpl (
     private val lastFMAPI: LastFMAPI,
     private val lastFMtoArtistResolver: LastFMToArtistResolver
-        ):ArtistExternalService{
+        ): ArtistExternalService {
 
      override fun  getArtistFromLastFMAPI(artistName: String): Artist.ArtistData? {
         val callResponse = lastFMAPI.getArtistInfo(artistName).execute()
