--- app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/model/entities/repository/externalArtist/ArtistInjector.kt
@@ -1,7 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.fulllogic.model.entities.repository.externalArtist
-
-import ayds.lisboa.songinfo.moredetails.fulllogic.model.entities.repository.externalArtist.artists.ArtistExternalInjector
-
-object ArtistInjector {
-    val artistExternalService : ArtistExternalService = ArtistExternalInjector.artistExternalService
-}
