--- app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/model/entities/repository/localArtist/ArtistLocalStorage.kt
@@ -1,9 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.fulllogic.model.entities.repository.localArtist
-
-import ayds.lisboa.songinfo.moredetails.fulllogic.model.entities.Artist
-
-interface ArtistLocalStorage {
-    fun getArtist(artistName: String): Artist
-
-    fun saveArtistInfo(artist: Artist?)
-}
